feat: added magic-router into tests
