+++ anchor-counter/package.json
@@ -5,7 +5,7 @@
   },
   "dependencies": {
     "@coral-xyz/anchor": "0.31.1",
-    "@magicblock-labs/ephemeral-rollups-sdk": "^0.2.4"
+    "@magicblock-labs/ephemeral-rollups-sdk": "^0.3.1"
   },
   "devDependencies": {
     "@types/bn.js": "^5.1.0",
