+++ anchor-counter/tests/anchor-counter.ts
@@ -9,6 +9,8 @@ import {
 const SEED_TEST_PDA = "test-pda"; // 5RgeA5P8bRaynJovch3zQURfJxXL3QK2JYg1YamSvyLb
 
 describe("anchor-counter", () => {
+  console.log("anchor-counter.ts")
+
   // Configure the client to use the local cluster.
   const provider = anchor.AnchorProvider.env();
   anchor.setProvider(provider);
@@ -38,7 +40,7 @@ describe("anchor-counter", () => {
 
   it("Initialize counter on Solana", async () => {
     const start = Date.now();
-    const txHash = await program.methods
+    let tx = await program.methods
       .initialize()
       .accounts({
         // @ts-ignore
@@ -46,7 +48,16 @@ describe("anchor-counter", () => {
         user: provider.wallet.publicKey,
         systemProgram: anchor.web3.SystemProgram.programId,
       })
-      .rpc({ skipPreflight: true });
+      .transaction();
+    tx.feePayer = provider.wallet.publicKey;
+    tx.recentBlockhash = (
+      await provider.connection.getLatestBlockhash()
+    ).blockhash;
+    tx = await providerEphemeralRollup.wallet.signTransaction(tx);
+    const txHash = await provider.sendAndConfirm(tx, [], {
+      skipPreflight: true,
+      commitment: "confirmed",
+    });
     const duration = Date.now() - start;
     console.log(`${duration}ms (Base Layer) Initialize txHash: ${txHash}`);
 
@@ -54,12 +65,21 @@ describe("anchor-counter", () => {
 
   it("Increase counter on Solana", async () => {
     const start = Date.now();
-    const txHash = await program.methods
+    let tx = await program.methods
       .increment()
       .accounts({
         counter: pda,
       })
-      .rpc();
+      .transaction();
+    tx.feePayer = provider.wallet.publicKey;
+    tx.recentBlockhash = (
+      await provider.connection.getLatestBlockhash()
+    ).blockhash;
+    tx = await providerEphemeralRollup.wallet.signTransaction(tx);
+    const txHash = await provider.sendAndConfirm(tx, [], {
+      skipPreflight: true,
+      commitment: "confirmed",
+    });
     const duration = Date.now() - start;
     console.log(`${duration}ms (Base Layer) Increment txHash: ${txHash}`);
   });
@@ -174,138 +194,6 @@ describe("anchor-counter", () => {
     const txHash = await providerEphemeralRollup.sendAndConfirm(tx);
     const duration = Date.now() - start;
     console.log(`${duration}ms (ER) Undelegate txHash: ${txHash}`);
-  });
-
-});
-
-
-describe("anchor-counter-increment-commit-atomic", () => {
-
-  // Configure the client to use the local cluster.
-  const provider = anchor.AnchorProvider.env();
-  anchor.setProvider(provider);
-
-  const providerEphemeralRollup = new anchor.AnchorProvider(
-    new anchor.web3.Connection(
-      process.env.EPHEMERAL_PROVIDER_ENDPOINT || "https://devnet-as.magicblock.app/",
-      {
-        wsEndpoint: process.env.EPHEMERAL_WS_ENDPOINT || "wss://devnet-as.magicblock.app/",
-      }
-    ),
-    anchor.Wallet.local()
-  );
-
-  before(async function () {
-    const balance = await provider.connection.getBalance(anchor.Wallet.local().publicKey)
-    console.log('Current balance is', balance / LAMPORTS_PER_SOL, ' SOL','\n')
-  })
-
-  const program = anchor.workspace.AnchorCounter as Program<AnchorCounter>;
-  const [pda] = anchor.web3.PublicKey.findProgramAddressSync(
-    [Buffer.from(SEED_TEST_PDA)],
-    program.programId
-  );
-
-  it("Initialize counter on Solana", async () => {
-    const start = Date.now();
-    const txHash = await program.methods
-      .initialize()
-      .accounts({
-        // @ts-ignore
-        counter: pda,
-        user: provider.wallet.publicKey,
-        systemProgram: anchor.web3.SystemProgram.programId,
-      })
-      .rpc({ skipPreflight: true });
-    const duration = Date.now() - start;
-    console.log(`${duration}ms (Base Layer) Initialize txHash: ${txHash}`);
-
-  });
-
-  it("Delegate counter to ER", async () => {
-    const start = Date.now();
-    let tx = await program.methods
-      .delegate()
-      .accounts({
-        payer: provider.wallet.publicKey,
-        pda: pda,
-      })
-      .transaction();
-    tx.feePayer = provider.wallet.publicKey;
-    tx.recentBlockhash = (
-      await provider.connection.getLatestBlockhash()
-    ).blockhash;
-    tx = await providerEphemeralRollup.wallet.signTransaction(tx);
-    const txHash = await provider.sendAndConfirm(tx, [], {
-      skipPreflight: true,
-      commitment: "confirmed",
-    });
-    const duration = Date.now() - start;
-    console.log(`${duration}ms (Base Layer) Delegate txHash: ${txHash}`);
-  });
-
-  it("Increase the delegate counter and commit through CPI", async () => {
-    const start = Date.now();
-    let tx = await program.methods
-      .incrementAndCommit()
-      .accounts({
-        payer: providerEphemeralRollup.wallet.publicKey,
-        // @ts-ignore
-        counter: pda,
-      })
-      .transaction();
-    tx.feePayer = providerEphemeralRollup.wallet.publicKey;
-    tx.recentBlockhash = (
-      await providerEphemeralRollup.connection.getLatestBlockhash()
-    ).blockhash;
-    tx = await providerEphemeralRollup.wallet.signTransaction(tx);
-
-    const txHash = await providerEphemeralRollup.sendAndConfirm(tx, [], {
-      skipPreflight: true,
-    });
-    const duration = Date.now() - start;
-    console.log(`${duration}ms (ER) Increment And Commit txHash: ${txHash}`);
-
-    // Get the commitment signature on the base layer
-    const comfirmCommitStart = Date.now();
-    // Await for the commitment on the base layer
-    const txCommitSgn = await GetCommitmentSignature(
-      txHash,
-      providerEphemeralRollup.connection
-    );
-    const commitDuration = Date.now() - comfirmCommitStart;
-    console.log(`${commitDuration}ms (Base Layer) Commit txHash: ${txCommitSgn}`);
-  });
-
-  it("Increase the delegate counter and undelegate through CPI", async () => {
-    const start = Date.now();
-    let tx = await program.methods
-      .incrementAndUndelegate()
-      .accounts({
-        payer: providerEphemeralRollup.wallet.publicKey,
-        // @ts-ignore
-        counter: pda,
-      })
-      .transaction();
-    tx.feePayer = provider.wallet.publicKey;
-    tx.recentBlockhash = (
-      await providerEphemeralRollup.connection.getLatestBlockhash()
-    ).blockhash;
-    tx = await providerEphemeralRollup.wallet.signTransaction(tx);
-
-    const txHash = await providerEphemeralRollup.sendAndConfirm(tx);
-    const duration = Date.now() - start;
-    console.log(`${duration}ms (ER) Increment and Undelegate txHash: ${txHash}`);
-
-    // Get the undelegate signature on the base layer
-    const comfirmCommitStart = Date.now();
-    // Await for the commitment on the base layer
-    const txCommitSgn = await GetCommitmentSignature(
-        txHash,
-        providerEphemeralRollup.connection
-    );
-    const commitDuration = Date.now() - comfirmCommitStart;
-    console.log(`${commitDuration}ms (Base Layer) Undelegate txHash: ${txCommitSgn}`);
   });
 
 });
